Add validation just before execution of `deploy` and `push` commands
