+++ cmd/docker-app/root.go
@@ -6,8 +6,6 @@ import (
 	"github.com/docker/app/internal"
 	"github.com/docker/cli/cli"
 	"github.com/docker/cli/cli/command"
-	cliconfig "github.com/docker/cli/cli/config"
-	"github.com/docker/cli/cli/debug"
 	cliflags "github.com/docker/cli/cli/flags"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
@@ -64,24 +62,5 @@ func addCommands(cmd *cobra.Command, dockerCli command.Cli) {
 			pullCmd(),
 			unpackCmd(),
 		)
-	}
-}
-
-func firstOrEmpty(list []string) string {
-	if len(list) != 0 {
-		return list[0]
-	}
-	return ""
-}
-
-func dockerPreRun(opts *cliflags.ClientOptions) {
-	cliflags.SetLogLevel(opts.Common.LogLevel)
-
-	if opts.ConfigDir != "" {
-		cliconfig.SetDir(opts.ConfigDir)
-	}
-
-	if opts.Common.Debug {
-		debug.Enable()
 	}
 }
